+++ SuperMarioBros-3/Effect.h
@@ -4,12 +4,12 @@
 class CEffect : public CGameObject
 {
 protected:
-	DWORD startEffect;
+	ULONGLONG startEffect;
 	int effectTime;
 public:
 	CEffect();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* objects);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* objects);
 	virtual void Render() = 0;
 	~CEffect();
 };
