+++ SuperMarioBros-3/GameObject.h
@@ -67,7 +67,7 @@ public:
 
 	int state;
 
-	DWORD dt; 
+	ULONGLONG dt; 
 
 	LPANIMATION_SET animation_set;
 
@@ -105,7 +105,7 @@ public:
 	CGameObject();
 
 	virtual void GetBoundingBox(float &left, float &top, float &right, float &bottom) = 0;
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT> *coObjects = NULL);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT> *coObjects = NULL);
 	virtual void Render() = 0;
 	virtual void SetState(int state) { this->state = state; }
 	void ResetAni();
