+++ SuperMarioBros-3/MoneyEffect.h
@@ -5,7 +5,7 @@ class CMoneyEffect : public CEffect
 {
 public:
 	CMoneyEffect(D3DXVECTOR2 position, int moneyAni = MONEY_EFFECT_100);
-	void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
+	void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects);
 	void Render();
 	~CMoneyEffect();
 };
