+++ SuperMarioBros-3/Tail.cpp
@@ -1,6 +1,6 @@
 #include "Tail.h"
 
-void CTail::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
+void CTail::Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects)
 {
 	for (LPGAMEOBJECT effect : listEffect)
 	{
